functional softmax. need to debug accuracy

+++ examples/softmax.py
@@ -6,7 +6,6 @@ import sys
 from itertools import product
 
 import numpy as np
-from neuronpy.core.language import bfloat16
 
 from autotune.cache.visualize import plot_metric
 from autotune.core.utils import GEMMCompatibility
@@ -19,7 +18,7 @@ from kernel_library.softmax import softmax_gemm_correctness_postprocessing
 
 def create_jobs(jobs: ProfileJobs, M: int, N: int, K: int):
     batch = 1
-    data_type = bfloat16
+    data_type = np.float32
     lhs = np.random.normal(size=(batch, M, K)).astype(data_type)
     rhs = np.random.normal(size=(K, N)).astype(data_type)
